+++ packages/adapters/wagmi/src/connectors/AuthConnector.ts
@@ -109,6 +109,7 @@ export function authConnector(parameters: AuthParameters) {
 
     async isAuthorized() {
       const provider = await this.getProvider()
+      console.log('>>> isAuthorized', provider.getLoginEmailUsed())
 
       return Promise.resolve(provider.getLoginEmailUsed())
     },