+++ packages/appkit/exports/constants.ts
@@ -1,1 +1,1 @@
-export const PACKAGE_VERSION = '1.5.1'
+export const PACKAGE_VERSION = '1.5.3-4228d6.0'