+++ packages/adapters/solana/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @reown/appkit-adapter-solana
 
+## 1.5.3-ae9267.0
+
+### Patch Changes
+
+- Sync wagmi status with account controller status
+
+- Updated dependencies []:
+  - @reown/appkit@1.5.3-ae9267.0
+  - @reown/appkit-common@1.5.3-ae9267.0
+  - @reown/appkit-core@1.5.3-ae9267.0
+  - @reown/appkit-utils@1.5.3-ae9267.0
+  - @reown/appkit-polyfills@1.5.3-ae9267.0
+  - @reown/appkit-scaffold-ui@1.5.3-ae9267.0
+  - @reown/appkit-ui@1.5.3-ae9267.0
+  - @reown/appkit-wallet@1.5.3-ae9267.0
+
 ## 1.5.2
 
 ### Patch Changes