+++ packages/core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @reown/appkit-core
 
+## 1.5.3-ae9267.0
+
+### Patch Changes
+
+- Sync wagmi status with account controller status
+
+- Updated dependencies []:
+  - @reown/appkit-common@1.5.3-ae9267.0
+  - @reown/appkit-wallet@1.5.3-ae9267.0
+
 ## 1.5.2
 
 ### Patch Changes